refactor: Move test files to test folder
